update fitur & konfigurasi

+++ src/app/api/contact/[id]/route.ts
@@ -1,14 +1,20 @@
 import { NextResponse } from 'next/server';
 import { prisma } from '@/lib/prisma';
 
-export async function DELETE(
-  request: Request,
-  { params }: { params: { id: string } }
-) {
+export async function DELETE(request: Request) {
   try {
+    // Ambil ID dari URL request
+    const { searchParams } = new URL(request.url);
+    const id = searchParams.get('id');
+
+    if (!id) {
+      return NextResponse.json({ error: 'Missing ID' }, { status: 400 });
+    }
+
     await prisma.visitor.delete({
-      where: { id: params.id }
+      where: { id }
     });
+
     return NextResponse.json({ message: 'Message deleted successfully' });
   } catch (error) {
     console.error('Error deleting message:', error);
@@ -17,4 +23,4 @@ export async function DELETE(
       { status: 500 }
     );
   }
-} 
+}
